refactor: add helpers
